requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
cryptography==41.0.7
numpy==1.24.3
pandas==2.1.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
asyncio==3.4.3
aioredis==2.0.1
motor==3.3.2
walrus-sdk==0.9.0
pysui==0.92.0